`int()` after rounding y_pred values

+++ src/pytorch/train_workflow.py
@@ -259,5 +259,5 @@ class TrainWorkflow:
                 pred_h = prefix_to_h(pred[i].tolist())
                 y_pred_values[x_str] = (y_h, pred_h)
             else:  # Regression
-                y_pred_values[x_str] = (torch.round(y[i][0]), torch.round(pred[i][0]))
+                y_pred_values[x_str] = (int(torch.round(y[i][0])), int(torch.round(pred[i][0])))
         return y_pred_values
